Fix compilation: remove WS2811_ALT_PINS, use pinList constructor, fix SD mode type, fix init order warnings, fix signedness, add MODE_TEST to switch

+++ UzomaBox/SDManager.cpp
@@ -42,7 +42,7 @@ void sdCardSkip(unsigned int len)
 }
 
 // ---- File open -------------------------------------------------------------
-bool sdFileOpen(const char *filename, uint8_t mode)
+bool sdFileOpen(const char *filename, int mode)
 {
   if (g_opened) sdFileClose();
   g_file = SD.open(filename, mode);
